fix(gtk): close gap between action button and ctx menu (#434 follow-up)
The first cut rounded trigger_height to whole cells, which left ~8px
of empty space below GTK's tab row (1.6 line_heights rounded up to 2).
Switch trigger_height from u16 to f32 (line_height units) so the GTK
caller can pass the exact 1.6 ratio and the menu sits flush against
the button bottom. TUI stays at 1.0 (one row).

+++ src/core/engine/mod.rs
@@ -1685,11 +1685,12 @@ pub struct ContextMenuState {
     pub selected: usize,
     pub screen_x: u16,
     pub screen_y: u16,
-    /// Height of the trigger element in cells. Non-zero values opt the
-    /// menu into `ContextMenuPlacement::Below` so it opens just below
-    /// the trigger instead of at the click point (#434). Right-click
-    /// flows leave this at 0 and use `AnchorPoint`.
-    pub trigger_height: u16,
+    /// Height of the trigger element in line_height units (f32 so
+    /// backends with sub-cell row heights — e.g. GTK's tab row at
+    /// ceil(1.6 * line_height) — can place the menu flush against the
+    /// trigger). Non-zero opts into `ContextMenuPlacement::Below`
+    /// (#434); right-click flows leave this at 0.0 and use AnchorPoint.
+    pub trigger_height: f32,
 }
 
 /// One visible row in the flat explorer file-tree list.
